Add certificate name checking to client config
This is because otherwise after gnome networkmanager import openvpn did not any
any checking on the certificate, other than the correct CA, because I think
"ns-cert-type server" parameter in the file is ignored in the importer.

This may open up a possible MITM attack, because if the attacker get hold of a
client certificate, it could be used as a fake server certificate. Not that
anyone would ever do that for VPN servers set up with this script...